IM7 (#261): LocalBuildBackend (Apple Containers > Docker BuildKit)
Second Phase 1 story for Epic IM. Brings up the build pipeline:
detect engine → render Dockerfile from TemplateSpec → stage build
context → invoke engine as child process → stream progress events
→ return BuildArtifact whose LocalReference IM6's LocalZotAdapter
can push to zot.

TemplateSpec extension (Andy.Containers/Abstractions/Images/):
- Added typed imperative fields: BaseImage, Extends, Packages, Files
  (TemplateFile record), Install, EntryPoint, Markers. Init-only props
  on the record so existing positional constructor calls keep working.
- TemplateFile record (Source, Dest, Mode?). Mode is Unix-permission
  octal in [0, 07777]; null means inherit COPY default.
- Documented the contract: SpecHash is computed against CanonicalJson,
  not these typed fields — they're a convenience view onto the same
  spec that the build backend consumes without re-parsing JSON.

Build infrastructure (Andy.Containers.Infrastructure/Build/):

- IBuildEngineDetector — interface; returns DetectedBuildEngine with
  Kind (None / AppleContainers / DockerBuildKit), ExecutablePath,
  ProbedVersion. Result cached for process lifetime.
- BuildEngineDetector — production impl. Probes `container --version`
  first (Apple priority per IM1), then `docker buildx version`
  (canonical buildkit-on-docker test — older docker installs lack
  buildx). 5s probe timeout. Caches via SemaphoreSlim. Logs the
  chosen engine at INFO; logs "no engine, builds will 503" at WARN
  when neither responds.
- DockerfileBuilder — pure static helper. Renders TemplateSpec to a
  Dockerfile. FROM/LABEL/RUN/COPY/ENTRYPOINT layout, one RUN per
  install command (granular cache invalidation), one COPY per file
  with chmod when Mode is set, package install with apt/dnf/apk
  picked from base-image heuristic, ENTRYPOINT in JSON-array exec
  form. Markers sorted lexicographically so identical specs produce
  identical bytes; that's the contract that makes spec-hash a
  meaningful cache key. Posix-style single-quote escaping for any
  value containing shell metacharacters.
- LocalBuildBackend — IBuildBackend impl. Renders Dockerfile via
  DockerfileBuilder, stages context in a temp directory (copies
  uploaded files under their LogicalName), invokes the detected
  engine via Process.Start with ProcessStartInfo.ArgumentList
  (defence-in-depth quoting). Streams stdout+stderr line-by-line as
  BuildStepStdoutEvent. On non-zero exit, throws
  ImageBuildFailedException with captured logs (consumed by IM10's
  422 mapping). On cancellation, kills the process tree and reports
  Cancelled outcome. Cleans up the build context unless
  PreserveBuildContext is set.
- LocalBuildBackendServiceCollectionExtensions — AddLocalBuildBackend()
  registers the detector + the backend (TryAddSingleton for both).

Tests (25 new, all passing):

- DockerfileBuilderTests (12): pure-function rendering rules.
  - FROM/LABEL/RUN/COPY/ENTRYPOINT layout
  - apt/dnf/apk package-manager detection (Debian, RHEL family,
    Alpine each pinned with theory data)
  - DEBIAN_FRONTEND=noninteractive + apt cache cleanup
  - One COPY per file, chmod only for files with explicit Mode
  - One RUN per install command (cache-invalidation contract)
  - ENTRYPOINT exec-form JSON array
  - Marker labels are deterministically sorted
  - Same input → same bytes (idempotency contract for spec hash)
  - Posix shell-metacharacter quoting (paths with spaces)
  - Throws when BaseImage missing (orchestrator must resolve extends
    chain before invoking the backend).

- BuildEngineDetectorTests (5): bash stubs simulate engines.
  - Apple Containers wins when both available (IM1 priority)
  - Falls back to Docker BuildKit when Apple absent
  - None when neither responds
  - Non-zero exit treated as engine-absent (not a fatal error)
  - Result cached: stub binary invoked exactly once across many
    DetectAsync calls.

- LocalBuildBackendTests (5): Moq IBuildEngineDetector + bash stub
  for the engine.
  - No-engine case throws ImageBuildFailedException with
    failingStepName=engine-detect and the actionable
    "install Apple Containers / Docker Desktop" message (so IM10
    has the substance to map onto a 503).
  - Successful build: stages context, renders Dockerfile, invokes
    engine with the right argv (BuildKit: `buildx build --load -t
    {tag} -f Dockerfile .`; Apple: `build -t {tag} -f Dockerfile .`),
    captures stdout as BuildStepStdoutEvent, completes with
    BuildOutcome.Succeeded.
  - Non-zero exit throws ImageBuildFailedException with stderr
    captured into CapturedLogs; failure path also emits the
    Failed completion event.
  - Capabilities reports the host architecture.

Test infrastructure refactor:
- Moved StubScript from DockerCliUploaderTests.cs to
  tests/Andy.Containers.Tests/TestSupport/StubScript.cs so IM6's
  uploader tests and IM7's engine + backend tests can share it.
  Added a (exitCode, stdoutLine, stderr) constructor variant for
  detector probes that need to read a version line from stdout.

DI wiring is added but NOT yet called from Program.cs — that hook-up
lands in IM8 (the orchestrator that ties build → push → DB row).

Andy.Containers.Tests: 450 / 450 green.
Whole solution builds 0 warning / 0 error under TreatWarningsAsErrors.

Closes #261.
